Create the app fresh on server start
