+++ lib/mongo.js
@@ -5,7 +5,7 @@
  * ==================================================
  * Mongoose MongoDB connection
  * ==================================================
- * 
+ *
  */
 
 var mongoose = require('mongoose');
@@ -28,22 +28,13 @@ exports.connect = function connect(done) {
     }
   };
 
-  mongoose.connect(settings.mongo, opts);
+  mongoose.connect(settings.mongo, opts, done);
 
   db = mongoose.connection;
-
-  db.on('error', function (error) {
-    console.log('Error connecting to MongoDB server:');
-    console.log(error);
-    console.log(error.stack);
-    throw error;
-  });
 
   db.on('reconnected', function () {
     console.log('We have reconnected to the MongoDB server.');
   });
-
-  db.once('open', done);
 
   return db;
 };
