+++ lib/server.js
@@ -10,8 +10,8 @@ var http = require('http');
 var _ = require('lodash');
 var async = require('async');
 var express = require('express');
+var logfmt = require('logfmt');
 var MongoStore = require('connect-mongo')(express);
-var uuid = require('node-uuid');
 var s3 = require('connect-s3');
 var passport = require('passport');
 
@@ -20,147 +20,87 @@ var settings = require('../settings');
 var routes = require('./routes');
 var postgres = require('./postgres');
 var tasks = require('./tasks');
+var util = require('./util');
 
 // Basic app variables
 var server;
-var app = express();
 var db = null;
 
-// ID generator
-var idgen = uuid.v1;
+function createApp(db) {
+  var app = express();
 
-// Log some information when requests are first received.
-if (process.env.REQUEST_LOGGER) {
-  app.use(express.logger({
-    immediate: true,
-    format: process.env.REQUEST_LOGGER
-  }));
-}
+  // Log some information when requests are first received.
+  if (process.env.REQUEST_LOGGER) {
+    app.use(express.logger({
+      immediate: true,
+      format: process.env.REQUEST_LOGGER
+    }));
+  }
 
-// Specify logging through an environment variable, so we can log differently
-// locally, on dev/test Heroku apps, and on production Heroku apps.
-if (process.env.EXPRESS_LOGGER) {
-  app.use(express.logger(process.env.EXPRESS_LOGGER));
-}
+  // Specify logging through an environment variable, so we can log differently
+  // locally, on dev/test Heroku apps, and on production Heroku apps.
+  if (process.env.EXPRESS_LOGGER) {
+    app.use(express.logger(process.env.EXPRESS_LOGGER));
+  }
 
-// Have Express wrap the response in a function for JSONP requests
-// https://github.com/visionmedia/express/issues/664
-app.set('jsonp callback', true);
+  // Have Express wrap the response in a function for JSONP requests
+  // https://github.com/visionmedia/express/issues/664
+  app.set('jsonp callback', true);
 
-// Use a compact JSON representation
-app.set('json spaces', 0);
+  // Use a compact JSON representation
+  app.set('json spaces', 0);
 
-// Allow Heroku to handle SSL for us
-app.enable('trust proxy');
+  // Allow Heroku to handle SSL for us
+  app.enable('trust proxy');
 
-// Allows clients to simulate DELETE and PUT
-// (some clients don't support those verbs)
-// http://stackoverflow.com/questions/8378338/what-does-connect-js-methodoverride-do
-app.use(express.methodOverride());
+  // Allows clients to simulate DELETE and PUT
+  // (some clients don't support those verbs)
+  // http://stackoverflow.com/questions/8378338/what-does-connect-js-methodoverride-do
+  app.use(express.methodOverride());
 
-// IE 8 and 9 can't post application/json for cross-origin requests, so we
-// accept text/plain treat it as JSON.
-// TODO: if we need to accept text/plain in the future, then we need to adjust
-// this
-function textParser(req, res, next) {
-  if (req._body) { return next(); }
-  req.body = req.body || {};
+  // Skip parsing the body for requests that we'll proxy to the tileserver
+  app.use('/tiles', util.rawParser);
 
-  // For GET/HEAD, there's no body to parse.
-  if ('GET' === req.method || 'HEAD' === req.method) { return next(); }
+  // Actually have Express parse text/plain as JSON
+  app.use(util.textParser);
 
-  // Check for text/plain content type
-  var type = req.headers['content-type'];
-  if (type === undefined || 'text/plain' !== type.split(';')[0]) { return next(); }
+  // Default to text/plain if no content-type was provided.
+  app.use(function(req, res, next) {
+    if (req.body) { return next(); }
+    if ('GET' === req.method || 'HEAD' === req.method) { return next(); }
 
-  // Flag as parsed
-  req._body = true;
-
-  console.log('info at=text_parser event=received_text/plain');
-
-  // Parse text/plain as if it was JSON
-  var buf = '';
-  req.setEncoding('utf8');
-  req.on('data', function(chunk){
-    buf += chunk;
-  });
-  req.on('end', function(){
-    try {
-      if (!buf.length) {
-        req.body = {};
-      } else {
-        req.body = JSON.parse(buf);
-      }
+    if (!req.headers['content-type']) {
+      req.body = {};
+      util.textParser(req, res, next);
+    } else {
       next();
-    } catch (err) {
-      err.status = 400;
-      next(err);
     }
   });
-}
 
-function rawParser(req, res, next) {
-  if (req._body) { return next(); }
-  req.body = req.body || {};
+  app.use(express.json());
+  app.use(express.urlencoded());
 
-  // For GET/HEAD, there's no body to parse.
-  if ('GET' === req.method || 'HEAD' === req.method) { return next(); }
+  // Keep extensions, so we can properly guess MIME types later.
+  // TODO: We should get the MIME type from the upload headers.
+  app.use(express.multipart({
+    keepExtensions: true
+  }));
 
-  // Flag as parsed
-  req._body = true;
+  // Let's compress everything!
+  app.use(express.compress());
 
-  var bufs = [];
-  var len = 0;
-  req.on('data', function(chunk){
-    bufs.push(chunk);
-    len += chunk.length;
-  });
-  req.on('end', function(){
-    req.body = Buffer.concat(bufs, len);
+  // Add common headers
+  app.use(function(req, res, next) {
+    res.header("Access-Control-Allow-Origin", "*");
+    res.header("Access-Control-Allow-Headers", "Origin, X-Mime-Type, X-Requested-With, X-File-Name, Content-Type");
+    res.header('Cache-Control', 'max-age=0, must-revalidate');
     next();
   });
+
+  return app;
 }
 
-// Skip parsing the body for requests that we'll proxy to the tileserver
-app.use('/tiles', rawParser);
-
-// Actually have Express parse text/plain as JSON
-app.use(textParser);
-
-// Default to text/plain if no content-type was provided.
-app.use(function(req, res, next) {
-  if (req.body) { return next(); }
-  if ('GET' === req.method || 'HEAD' === req.method) { return next(); }
-
-  if (!req.headers['content-type']) {
-    req.body = {};
-    textParser(req, res, next);
-  } else {
-    next();
-  }
-});
-
-app.use(express.json());
-app.use(express.urlencoded());
-
-// Keep extensions, so we can properly guess MIME types later.
-// TODO: We should get the MIME type from the upload headers.
-app.use(express.multipart({
-  keepExtensions: true
-}));
-
-// Let's compress everything!
-app.use(express.compress());
-
-// Add common headers
-app.use(function(req, res, next) {
-  res.header("Access-Control-Allow-Origin", "*");
-  res.header("Access-Control-Allow-Headers", "Origin, X-Mime-Type, X-Requested-With, X-File-Name, Content-Type");
-  res.header('Cache-Control', 'max-age=0, must-revalidate');
-  next();
-});
-
-function setupAuth(db) {
+function setupAuth(app, db) {
   // Authentication-related middleware
   app.use(express.cookieParser());
 
@@ -187,7 +127,7 @@ function setupAuth(db) {
   app.set('session-middleware', session);
 }
 
-function setupRoutes() {
+function setupRoutes(app) {
   routes.setup(app);
 
   // Serve the mobile collection app from /mobile
@@ -226,7 +166,7 @@ function setupRoutes() {
 
 // We're done with our preflight stuff.
 // Now, we start listening for requests.
-function startServer(port, cb) {
+function startServer(app, port, cb) {
   server = http.createServer(app);
   server.listen(port, function (err) {
     console.log('info at=server event=listening port=' + port);
@@ -253,12 +193,21 @@ function run(cb) {
       cb = function () {};
     }
 
-    db = mongo.connect(function () {
-      setupAuth(db);
-      setupRoutes(); // Needs to happen AFTER setupAuth
+    db = mongo.connect(function (error) {
+      if (error) {
+        logfmt.error(error);
+        cb(error);
+        return;
+      }
+
+      logfmt.log({ level: 'info', at: 'server', event: 'mongo_connected' });
+
+      var app = createApp(db);
+      setupAuth(app, db);
+      setupRoutes(app); // Needs to happen AFTER setupAuth
       async.series([
         tasks.connect,
-        startServer.bind(null, settings.port)
+        startServer.bind(null, app, settings.port)
       ], cb);
     });
