+++ lib/util.js
@@ -288,3 +288,66 @@ util.simplifyCoordinates = (function () {
 
   return simplify;
 }());
+
+// IE 8 and 9 can't post application/json for cross-origin requests, so we
+// accept text/plain treat it as JSON.
+// TODO: if we need to accept text/plain in the future, then we need to adjust
+// this
+util.textParser = function textParser(req, res, next) {
+  if (req._body) { return next(); }
+  req.body = req.body || {};
+
+  // For GET/HEAD, there's no body to parse.
+  if ('GET' === req.method || 'HEAD' === req.method) { return next(); }
+
+  // Check for text/plain content type
+  var type = req.headers['content-type'];
+  if (type === undefined || 'text/plain' !== type.split(';')[0]) { return next(); }
+
+  // Flag as parsed
+  req._body = true;
+
+  console.log('info at=text_parser event=received_text/plain');
+
+  // Parse text/plain as if it was JSON
+  var buf = '';
+  req.setEncoding('utf8');
+  req.on('data', function(chunk){
+    buf += chunk;
+  });
+  req.on('end', function(){
+    try {
+      if (!buf.length) {
+        req.body = {};
+      } else {
+        req.body = JSON.parse(buf);
+      }
+      next();
+    } catch (err) {
+      err.status = 400;
+      next(err);
+    }
+  });
+};
+
+util.rawParser = function rawParser(req, res, next) {
+  if (req._body) { return next(); }
+  req.body = req.body || {};
+
+  // For GET/HEAD, there's no body to parse.
+  if ('GET' === req.method || 'HEAD' === req.method) { return next(); }
+
+  // Flag as parsed
+  req._body = true;
+
+  var bufs = [];
+  var len = 0;
+  req.on('data', function(chunk){
+    bufs.push(chunk);
+    len += chunk.length;
+  });
+  req.on('end', function(){
+    req.body = Buffer.concat(bufs, len);
+    next();
+  });
+};
